Renamed FixedSidebar -> Sidebar inorder to avoid errors in code
