+++ frontend/src/components/dashboard/Sidebar.tsx
